Changed Multiple Files
DatabaseManager - Changed an SQL datatype to store as bytes rather than as a string.
EncryptionMethods - Changed the encryption methods to return the IV of the data.
Logger - Fixed formating issues.
EncryptionMethodsTest - Added an extra test to ensure that it generates an IV for AES encryption if not provided one. Also updated the AES Encryption test to ensure it works properly.